hospital_management123/hospital_management/backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
pytorch==1.9.0
transformers==4.11.3
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
python-dotenv==0.19.0